fix: 테스트 코드 Mock 추가
- 변경된 알림 기능의 Mock 객체 필드 추가
- PaymentService 가격 데이터 타입 오류 수정

Refs: #94

+++ src/main/java/com/sparta/cream/service/PaymentService.java
@@ -166,7 +166,7 @@ public class PaymentService {
 			eventPublisher.publishEvent(new PaymentCompletedEvent(
 				userId,
 				payment.getProductName(),
-				payment.getAmount(),
+				payment.getAmount().longValue(),
 				payment.getId()
 			));
